backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
ldap3==2.9.1
pydantic==2.5.3
python-dotenv==1.0.0
pillow==10.1.0
psycopg2-binary==2.9.12
PyJWT==2.8.0